Added support to use the application with remote machines
In order to get this working I had to made some changes on the WMI calls, as well as the File Paths returned for the service.

Added a new private property to store the name of the server we'll be connecting to.

Modified the GetService()  call to request the information to the remote machine, and added wrapped it in a try/catch to catch access/connection issues.

Added new elements to the design to get input from the user for the server connection, and methods to handle connections, and disconnection events.

Added a statusstrip to easy to see show connection information, and mofied the title to show only Total: 0 if disconnected, as it would be weirdly formatted, 'Total: 0,', if it was disconnected.